Remove seperate counts for used and cached blocks in memory_arena
There is only one count for the capacity(), the size() is calculated in O(n).
This shrinks down the object size, O(n) size() isn't so bad since it is rarely used and usually small.
Making capacity() O(n) would be slower, since the capacity() is bigger.
If the cache is empty, the size is equivalent to the capacity(), so it can be O(1) then.