Change return type of allocator_reference get_allocator() to a reference in each case
The return value of a StoragePolicy's get_allocator() must now be a reference. This gets rid of the mess of reference vs value return types and makes lock() for all cases possible.

Also cleanup of const-correctness in allocator_storage, all references return non-const in all cases and allocator_storage's lock() reflects that.