Fix crash in detail::ordered_free_memory_list
When allocating arrays, last_dealloc_ wasn't properly updated if removing a sequence starting with the first node. This was because less(i.prev, last_dealloc_) doesn't work as intended when i.prev is the proxy.

Change comparison to less_equal with i.first which is never proxy.
Likewise for end comparison.

Also add less_equal()/greater_equal() for pointers.