Disable debug fences in free lists
This simplifes the implementation and avoids some potential alignment bugs (if there were any).
If needed, debug fences can be added back as part of the memory pool logic:
increase the node size there by the fence size; this wastes memory for arrays, but keeps the free list implementation clean.